refactor(*): refactor SDK and cleanup
