--- .github/workflows/lint.yml
@@ -1,18 +1,0 @@
-name: Lint
-on: [push, pull_request]
-jobs:
-  golangci:
-    name: lint
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        go-version: [1.17.x]
-    steps:
-      - uses: actions/setup-go@v3
-        with:
-          go-version: ${{ matrix.go-version }}
-      - uses: actions/checkout@v3
-      - name: golangci-lint
-        uses: golangci/golangci-lint-action@v3
-        with:
-          version: v1.43
