--- .golangci.yml
@@ -1,3 +1,0 @@
-run:
-  skip-dirs:
-    - analyzers/triggers
