+++ analyzers/testdata/src/staticcheck/staticcheck.go
